+++ dotnet-test-samples/hexagonal-architecture/src/GetStock/Adapters/CurrencyConverterHttpClient.cs
@@ -1,66 +1,9 @@
 ﻿using System.Text.Json;
-using System.Text.Json.Serialization;
 using Amazon.Lambda.Core;
+using GetStock.Adapters.Model;
 
 namespace GetStock.Adapters
 {
-    internal class CurrencyRates
-    {
-        [JsonPropertyName("base")]
-        public string Base { get; set; }
-
-        [JsonPropertyName("date")]
-        public string Date { get; set; }
-
-        [JsonPropertyName("rates")]
-        public Dictionary<string, double> Rates { get; set; }
-
-        [JsonPropertyName("success")]
-        public bool Success { get; set; }
-
-        [JsonPropertyName("timestamp")]
-        public int Timestamp { get; set; }
-    }
-
-    public interface IHttpClient : IDisposable
-    {
-        Task<string> GetAsync(string url);
-    }
-
-    internal class HttpClientWrapper : IHttpClient
-    {
-        private readonly HttpClient _httpClient;
-
-        public HttpClientWrapper(IServiceConfiguration configuration)
-        {
-            _httpClient = new HttpClient
-            {
-                BaseAddress = new Uri(configuration.CurrencyBaseAddress)
-            };
-
-            _httpClient.DefaultRequestHeaders.Add("apikey", configuration.CurrencyApiKey);
-        }
-
-
-        public async Task<string> GetAsync(string url)
-        {
-            var response = await _httpClient.GetAsync(url);
-            if (!response.IsSuccessStatusCode)
-            {
-                LambdaLogger.Log("HTTP call failed: " + response);
-                LambdaLogger.Log("URL: " + url);
-                return string.Empty;// TODO: exception?
-            }
-
-            return await response.Content.ReadAsStringAsync();
-        }
-
-        public void Dispose()
-        {
-            _httpClient?.Dispose();
-        }
-    }
-
     public class CurrencyConverterHttpClient : ICurrencyConverter, IDisposable
     {
         private readonly IHttpClient _httpHandler;
@@ -85,7 +28,7 @@ namespace GetStock.Adapters
             LambdaLogger.Log("response: " + responseJson);
 
             var result = JsonSerializer.Deserialize<CurrencyRates>(responseJson);
-            if (result == null || !result.Success)
+            if (result is not { Success: true })
             {
                 return new Dictionary<string, double>();
             }
