+++ dotnet-test-samples/hexagonal-architecture/src/GetStock/Adapters/IHttpClient.cs
@@ -1,0 +1,6 @@
+﻿namespace GetStock.Adapters;
+
+public interface IHttpClient : IDisposable
+{
+    Task<string> GetAsync(string url);
+}
