+++ dotnet-test-samples/hexagonal-architecture/src/GetStock/Domains/IStockLogic.cs
@@ -1,0 +1,8 @@
+﻿using GetStock.Domains.Models;
+
+namespace GetStock.Domains;
+
+public interface IStockLogic
+{
+    Task<StockWithCurrencies> RetrieveStockValuesAsync(string stockId);
+}
