+++ dotnet-test-samples/hexagonal-architecture/src/GetStock/Domains/StockLogic.cs
@@ -1,37 +1,11 @@
 ﻿using Amazon.Lambda.Core;
 using GetStock.Adapters.Exceptions;
+using GetStock.Domains.Models;
 using GetStock.Ports;
 using static GetStock.Utilities.CollectionUtils;
 
 namespace GetStock.Domains
 {
-    public readonly struct StockWithCurrencies
-    {
-        public string StockId { get; }
-
-        public IEnumerable<KeyValuePair<string, double>> Values { get; }
-
-        public StockWithCurrencies(string stockId, IEnumerable<KeyValuePair<string, double>> values)
-        {
-            StockId = stockId;
-            Values = values;
-        }
-
-        public override bool Equals(object? obj)
-        {
-            return obj is StockWithCurrencies currencies &&
-                   StockId == currencies.StockId &&
-                   Values.Count() == currencies.Values.Count() && (!Values.Except(currencies.Values).Any() || !currencies.Values.Except(Values).Any());
-        }
-
-        public override int GetHashCode() => HashCode.Combine(StockId, Values);
-    }
-
-    public interface IStockLogic
-    {
-        Task<StockWithCurrencies> RetrieveStockValuesAsync(string stockId);
-    }
-
     internal class StockLogic : IStockLogic
     {
         private const string BaseCurrency = "EUR";
